Renaming of Buffers to match texture pattern
