+++ CMakeLists.txt
@@ -115,15 +115,20 @@ set(${PROJECT_NAME}_SOURCE
     "src/common/BasicCamera.cpp"
     "src/common/FileTexture.cpp"
     "src/common/Handle.cpp"
-    "src/common/buffers/GlobalInfo.cpp"
-    "src/common/buffers/LightInfo.cpp"
-    "src/common/buffers/InstanceModelInfo.cpp"
-    "src/common/buffers/InstanceNormalInfo.cpp"
-    "src/common/buffers/ObjVertInfo.cpp"
-    "src/common/buffers/ObjIndicesInfo.cpp"
-    "src/common/buffers/ObjVertInfoFile.cpp"
+    "src/common/buffers/TransferRequest_GlobalInfo.cpp"
+    "src/common/buffers/TransferRequest_InstanceModelInfo.cpp"
+    "src/common/buffers/TransferRequest_InstanceNormalInfo.cpp"
+    "src/common/buffers/TransferRequest_LightInfo.cpp"
+    "src/common/buffers/TransferRequest_IndiciesInfo.cpp"
+    "src/common/buffers/TransferRequest_VertInfo.cpp"
     "src/common/textures/TransferRequest_TextureFile.cpp"
     "src/common/controllers/ManagerController_RenderResource_TextureFile.cpp"
+    "src/common/controllers/ManagerController_RenderResource_GlobalInfo.cpp"
+    "src/common/controllers/ManagerController_RenderResource_InstanceModelInfo.cpp"
+    "src/common/controllers/ManagerController_RenderResource_InstanceNormalInfo.cpp"
+    "src/common/controllers/ManagerController_RenderResource_LightInfo.cpp"
+    "src/common/controllers/ManagerController_RenderResource_IndicesInfo.cpp"
+    "src/common/controllers/ManagerController_RenderResource_VertInfo.cpp"
     "src/virtual/StarSystemPipeline.cpp"
     "src/common/materials/TextureMaterial.cpp"
     "src/common/helpers/GeometryHelpers.cpp"
@@ -150,15 +155,19 @@ set(${PROJECT_NAME}_HEADERS
     "include/common/objects/StarObjectInstance.hpp"
     "include/common/BasicCamera.hpp"
     "include/common/RuntimeUpdateTexture.hpp"
-    "include/common/buffers/GlobalInfo.hpp"
-    "include/common/buffers/LightInfo.hpp"
-    "include/common/buffers/InstanceModelInfo.hpp"
-    "include/common/buffers/InstanceNormalInfo.hpp"
-    "include/common/buffers/ObjVertInfo.hpp"
-    "include/common/buffers/ObjIndicesInfo.hpp"
-    "include/common/buffers/ObjVertInfoFile.hpp"
+    "include/common/buffers/TransferRequest_GlobalInfo.hpp"
+    "include/common/buffers/TransferRequest_InstanceModelInfo.hpp"
+    "include/common/buffers/TransferRequest_InstanceNormalInfo.hpp"
+    "include/common/buffers/TransferRequest_IndicesInfo.hpp"
+    "include/common/buffers/TransferRequest_VertInfo.hpp"
     "include/common/textures/TransferRequest_TextureFile.hpp"
     "include/common/controllers/ManagerController_RenderResource_TextureFile.hpp"
+    "include/common/controllers/ManagerController_RenderResource_GlobalInfo.hpp"
+    "include/common/controllers/ManagerController_RenderResource_InstanceModelInfo.hpp"
+    "include/common/controllers/ManagerController_RenderResource_InstanceNormalInfo.hpp"
+    "include/common/controllers/ManagerController_RenderResource_LightInfo.hpp"
+    "include/common/controllers/ManagerController_RenderResource_IndicesInfo.hpp"
+    "include/common/controllers/ManagerController_RenderResource_VertInfo.hpp"
     "include/virtual/ModulePlug/RenderResourceModifier.hpp"
     "include/virtual/ModulePlug/CommandBufferModifier.hpp"
     "include/virtual/ModulePlug/DescriptorModifier.hpp"
