--- include/common/buffers/ObjAABBInfo.hpp
@@ -1,15 +1,0 @@
-#pragma once
-
-#include "BufferManagerRequest.hpp"
-#include "BufferMemoryTransferRequest.hpp"
-
-namespace star{
-    class AABBInfoTransfer : public BufferMemoryTransferRequest{
-    };
-
-    class ObjBoundingBoxInfo : public RequestManagerRenderResource{
-
-    };
-
-
-}
