+++ include/virtual/ManagerController_RenderResource_Buffer.hpp
@@ -12,7 +12,7 @@ namespace star::ManagerController::RenderResource{
     Buffer(const uint8_t& frameInFlightIndexToUpdateOn) 
     : star::ManagerController::Controller<TransferRequest::Memory<StarBuffer::BufferCreationArgs>>(frameInFlightIndexToUpdateOn){}; 
 
-    virtual std::unique_ptr<TransferRequest::Memory<StarBuffer::BufferCreationArgs>> createTransferRequest() const override = 0;
+    virtual std::unique_ptr<TransferRequest::Memory<StarBuffer::BufferCreationArgs>> createTransferRequest() override = 0;
 
     protected:
