--- src/common/buffers/ObjAABBInfo.cpp
@@ -1,1 +1,0 @@
-#include "ObjAABBInfo.hpp"
