+++ src/common/controllers/ManagerController_RenderResource_TextureFile.cpp
@@ -2,6 +2,6 @@
 
 #include "TransferRequest_TextureFile.hpp"
 
-std::unique_ptr<star::TransferRequest::Memory<star::StarTexture::TextureCreateSettings>> star::ManagerController::RenderResource::TextureFile::createTransferRequest() const{
+std::unique_ptr<star::TransferRequest::Memory<star::StarTexture::TextureCreateSettings>> star::ManagerController::RenderResource::TextureFile::createTransferRequest(){
     return std::make_unique<TransferRequest::TextureFile>(this->filePath);
 }
