+++ src/virtual/StarScene.cpp
@@ -1,12 +1,15 @@
 #include "StarScene.hpp"
 
+#include "ManagerController_RenderResource_GlobalInfo.hpp"
+#include "ManagerController_RenderResource_LightInfo.hpp"
+
 star::StarScene::StarScene(const int& numFramesInFlight) 
 : myCamera(std::make_optional(std::make_unique<star::BasicCamera>(1280, 720))){
 
 	this->lightInfoBuffers.resize(numFramesInFlight);
 
 	for (int i = 0; i < numFramesInFlight; i++) {
-		this->globalInfoBuffers.emplace_back(ManagerRenderResource::addRequest(std::make_unique<GlobalInfo>(static_cast<uint16_t>(i), *this->myCamera.value(), this->lightCounter)));
+		this->globalInfoBuffers.emplace_back(ManagerRenderResource::addRequest(std::make_unique<ManagerController::RenderResource::GlobalInfo>(static_cast<uint16_t>(i), *this->myCamera.value(), this->lightCounter)));
 	}
 }
 
@@ -26,16 +29,9 @@ int star::StarScene::add(std::unique_ptr<star::Light> newLight) {
 	for (int i = 0; i < this->lightInfoBuffers.size(); i++) {
 		//TODO: need a function which will replace a buffer contained within the buffer manager...all handles to the information represented in that buffer need to remain valid
 		if (!this->lightInfoBuffers[i].isInitialized())
-			this->lightInfoBuffers[i] = ManagerRenderResource::addRequest(std::make_unique<LightInfo>(i, this->lightList));
-		else
-			ManagerRenderResource::updateRequest(std::make_unique<LightInfo>(i, this->lightList), this->lightInfoBuffers[i]); 
+			this->lightInfoBuffers[i] = ManagerRenderResource::addRequest(std::make_unique<ManagerController::RenderResource::LightInfo>(i, this->lightList));
 	}
 
-	for (int i = 0; i < this->globalInfoBuffers.size(); i++){
-		// if (this->myCamera.has)
-
-		// ManagerBuffer::updateRequest(std::make_unique<GlobalInfo>(i, *this->camera, this->lightCounter), this->globalInfoBuffers[i]); 
-	}
 	return lightIndex;
 }
 
@@ -59,6 +55,6 @@ void star::StarScene::onWorldUpdate(const uint32_t& frameInFlightIndex){
 	//check if any objects have changed since last frame, if so, they need to be updated on GPU memory
 
 	if (this->myCamera.has_value()){
-		ManagerRenderResource::updateRequest(std::make_unique<GlobalInfo>(frameInFlightIndex, *this->myCamera.value(), this->lightList.size()), this->globalInfoBuffers[frameInFlightIndex]);
+		ManagerRenderResource::updateRequest(std::make_unique<ManagerController::RenderResource::GlobalInfo>(frameInFlightIndex, *this->myCamera.value(), this->lightList.size()), this->globalInfoBuffers[frameInFlightIndex]);
 	}
 }
